Optimize sequential word progression and add streak calendar
